requirements/base.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy[asyncio]
alembic
aiosqlite
asyncpg
aiomysql

# Redis and Celery
redis[hiredis]
celery[redis]

# AI and HTTP
httpx
ollama

# Data validation
pydantic
pydantic-settings

# Utilities
python-dotenv
PyYAML
itsdangerous